{"types":"C","runs":25}

+++ cons.java
@@ -8,7 +8,7 @@ public class cons extends ConsoleProgram {
 		int c = readInt("enter third number : ");
 		//println((a + b + c)/3);
 		
-		double d = (double)((1 / a) + (1 / b) + (1 / c)) ;
+		double d = (double)((1 /(double) a) + (1 / (double) b) + (1 / (double)c)) ;
 		//System.out.print(1/a);
 		println( 1 / d );
